django-starter-jwt/requirements.txt
asgiref==3.3.1
Django==3.1.5
django-cors-middleware==1.5.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==2.0.1
pytz==2020.5
sqlparse==0.4.1
